fix: use env.APP_TEMP_DIR instead of KysoSettings
This commit removes `REPORT_PATH` and `TMP_FOLDER_PATH` from the
`KysoSettings` collection and uses the environment variable `APP_TEMP_DIR`
instead.

If the variable is not set the api uses the value
`${process.pwd()}/data` for it and, in any case, it tries to create the
folder if it does not exist, terminating the process if it can't be
created.